Move source-related sidebar text constants to top-level constants file
Now all personal dashboard sidebar copy is in one place
