+++ x-pack/plugins/enterprise_search/public/applications/workplace_search/components/layout/personal_dashboard_sidebar/private_sources_sidebar.test.tsx
@@ -11,14 +11,14 @@ import React from 'react';
 
 import { shallow } from 'enzyme';
 
-import { SourceSubNav } from '../../../views/content_sources/components/source_sub_nav';
-
 import {
   PRIVATE_CAN_CREATE_PAGE_TITLE,
   PRIVATE_VIEW_ONLY_PAGE_TITLE,
   PRIVATE_VIEW_ONLY_PAGE_DESCRIPTION,
   PRIVATE_CAN_CREATE_PAGE_DESCRIPTION,
-} from '../../../views/content_sources/constants';
+} from '../../../constants';
+import { SourceSubNav } from '../../../views/content_sources/components/source_sub_nav';
+
 import { ViewContentHeader } from '../../shared/view_content_header';
 
 import { PrivateSourcesSidebar } from './private_sources_sidebar';
